Added general dialog box + browser icons
Created some CSS for a general dialog box with an overlay behind it.
The box itself and paragraphs within it is styled so far, but general
links, buttons, and form inputs need to be styled for it to be able to
be used consistently. The Server box and Authoring Mode login box can
now be redesigned with this CSS.

For the incompatible browser dialog box:
- Added browser icons